code/load/requirements.txt
watchdog
matplotlib
numpy
pandas
rdflib
mysql-connector-python
docker
sparqlwrapper